pyproject.toml
[tool.poetry]
name = "yundownload"
version = "0.6.21"
description = "file downloader"
authors = ["yunhai <bybxbwg@foxmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.28.1"
aiofiles = "^24.1.0"
paramiko = "^3.5.1"
m3u8 = "^6.0.0"
colorlog = "^6.9.0"
pycryptodome = "^3.23.0"

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.43"
mkdocs-git-revision-date-localized-plugin = "^1.3.0"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yundownload = "yundownload:cli"
